requirements.txt
openai==1.12.0
python-dotenv==1.0.1
python-telegram-bot==20.8